+++ R/data_shifted_peaks.R
@@ -1,0 +1,9 @@
+#' "Shifted Peaks" Simulated Dataset
+#'
+#' A simulated dataset generated for examples in the veesa pipeline manuscript. 
+#' For the code used to prepare this dataset, see
+#' https://github.com/sandialabs/veesa/inst/data-shifted-peaks.md.
+#'
+#' @format A data.frame.
+
+"shifted_peaks"