+++ NEWS.md
@@ -1,3 +1,8 @@
+# veesa 0.1.2
+
+- Added shifted peaks data
+- Added example code from manuscript
+
 # veesa 0.1.1
 
 - Adjusted title in plot_pc_directions